Update atom visualization
The project "Graphical Visualization of Atom" is a computer graphics visualization application developed using CGV (Computer Graphics and Visualization) techniques. The
aim of this project is to provide an interactive and visually appealing representation of atoms, allowing users to explore and understand their structure.
The application utilizes advanced graphics algorithms and techniques to create a realistic and dynamic visualization of atoms. It employs techniques such as particle systems, shaders, and lighting to accurately depict the behavior and properties of atoms. The project incorporates scientific data and models to ensure accuracy in the visual representation.